[table-driven-branch] Use test helpers again.
Previously I had to copy the encode/decode helpers into each test
file so that they could call the generated concrete type's methods
instead of the ones from the protocol extension (which use the
visitor/decoder pattern). For now, I've added a few more temporary
requirements to `Message` so that we can make them work in both
cases. Now the test cases can use the generic helpers again, cutting
out some annoying repetition.
