+++ Tests/ExperimentalTableDrivenSwiftProtobufTests/Test_JSON.swift
@@ -16,7 +16,7 @@ import Foundation
 import SwiftProtobuf
 import XCTest
 
-final class Test_JSON: XCTestCase {
+final class Test_JSON: XCTestCase, PBTestHelpers {
     typealias MessageTestType = SwiftProtoTesting_TestAllTypes
 
     func testOptionalInt32() {
@@ -775,178 +775,6 @@ final class Test_JSON: XCTestCase {
         assertJSONDecodeFails("")
         assertJSONDecodeFails("{")
         assertJSONDecodeFails("}")
-    }
-
-    func assertJSONEncode(
-        _ expected: String,
-        extensions: any ExtensionMap = SimpleExtensionMap(),
-        encodingOptions: JSONEncodingOptions = .init(),
-        file: StaticString = #file,
-        line: UInt = #line,
-        configure: (inout MessageTestType) -> Void
-    ) {
-        let empty = MessageTestType()
-        var configured = empty
-        configure(&configured)
-        XCTAssert(configured != empty, "Object should not be equal to empty object", file: file, line: line)
-        do {
-            let encoded = try configured.jsonString(options: encodingOptions)
-            XCTAssert(
-                expected == encoded,
-                "Did not encode correctly: got \(encoded) but expected \(expected)",
-                file: file,
-                line: line
-            )
-            do {
-                let decoded = try MessageTestType(jsonString: encoded, extensions: extensions)
-                XCTAssert(
-                    decoded == configured,
-                    "Encode/decode cycle should generate equal object: \(decoded) != \(configured)",
-                    file: file,
-                    line: line
-                )
-            } catch {
-                XCTFail(
-                    "Encode/decode cycle should not throw error decoding: \(encoded), but it threw \(error)",
-                    file: file,
-                    line: line
-                )
-            }
-        } catch let e {
-            XCTFail("Failed to serialize JSON: \(e)\n    \(configured)", file: file, line: line)
-        }
-
-        do {
-            let encodedData: [UInt8] = try configured.jsonUTF8Bytes(options: encodingOptions)
-            let encodedOptString = String(bytes: encodedData, encoding: String.Encoding.utf8)
-            XCTAssertNotNil(encodedOptString)
-            let encodedString = encodedOptString!
-            XCTAssert(
-                expected == encodedString,
-                "Did not encode correctly: got \(encodedString)",
-                file: file,
-                line: line
-            )
-            do {
-                let decoded = try MessageTestType(jsonUTF8Bytes: encodedData, extensions: extensions)
-                XCTAssert(
-                    decoded == configured,
-                    "Encode/decode cycle should generate equal object: \(decoded) != \(configured)",
-                    file: file,
-                    line: line
-                )
-            } catch {
-                XCTFail(
-                    "Encode/decode cycle should not throw error decoding: \(encodedString), but it threw \(error)",
-                    file: file,
-                    line: line
-                )
-            }
-        } catch let e {
-            XCTFail("Failed to serialize JSON: \(e)\n    \(configured)", file: file, line: line)
-        }
-    }
-
-    func assertJSONDecodeSucceeds(
-        _ json: String,
-        options: JSONDecodingOptions = JSONDecodingOptions(),
-        extensions: any ExtensionMap = SimpleExtensionMap(),
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) -> Bool
-    ) {
-        do {
-            let decoded: MessageTestType = try MessageTestType(
-                jsonString: json,
-                extensions: extensions,
-                options: options
-            )
-            XCTAssert(check(decoded), "Condition failed for \(decoded)", file: file, line: line)
-
-            do {
-                let encoded = try decoded.jsonString()
-                do {
-                    let redecoded = try MessageTestType(jsonString: encoded, extensions: extensions, options: options)
-                    XCTAssert(
-                        check(redecoded),
-                        "Condition failed for redecoded \(redecoded) from \(encoded)",
-                        file: file,
-                        line: line
-                    )
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch {
-                    XCTFail("Swift should have recoded/redecoded without error: \(encoded)", file: file, line: line)
-                }
-            } catch let e {
-                XCTFail("Swift should have recoded without error but got \(e)\n    \(decoded)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Swift should have decoded without error but got \(e): \(json)", file: file, line: line)
-            return
-        }
-
-        do {
-            let jsonData = json.data(using: String.Encoding.utf8)!
-            let decoded: MessageTestType = try MessageTestType(
-                jsonUTF8Bytes: jsonData,
-                extensions: extensions,
-                options: options
-            )
-            XCTAssert(check(decoded), "Condition failed for \(decoded) from binary \(json)", file: file, line: line)
-
-            do {
-                let encoded: [UInt8] = try decoded.jsonUTF8Bytes()
-                let encodedString = String(decoding: encoded, as: UTF8.self)
-                do {
-                    let redecoded = try MessageTestType(
-                        jsonUTF8Bytes: encoded,
-                        extensions: extensions,
-                        options: options
-                    )
-                    XCTAssert(
-                        check(redecoded),
-                        "Condition failed for redecoded \(redecoded) from binary \(encodedString)",
-                        file: file,
-                        line: line
-                    )
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch {
-                    XCTFail(
-                        "Swift should have recoded/redecoded without error: \(encodedString)",
-                        file: file,
-                        line: line
-                    )
-                }
-            } catch let e {
-                XCTFail("Swift should have recoded without error but got \(e)\n    \(decoded)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Swift should have decoded without error but got \(e): \(json)", file: file, line: line)
-            return
-        }
-    }
-
-    func assertJSONDecodeFails(
-        _ json: String,
-        extensions: any ExtensionMap = SimpleExtensionMap(),
-        options: JSONDecodingOptions = JSONDecodingOptions(),
-        file: StaticString = #file,
-        line: UInt = #line
-    ) {
-        do {
-            let _ = try MessageTestType(jsonString: json, extensions: extensions, options: options)
-            XCTFail("Swift decode should have failed: \(json)", file: file, line: line)
-        } catch {
-            // Yay! It failed!
-        }
-
-        let jsonData = json.data(using: String.Encoding.utf8)!
-        do {
-            let _ = try MessageTestType(jsonUTF8Bytes: jsonData, extensions: extensions, options: options)
-            XCTFail("Swift decode should have failed for binary: \(json)", file: file, line: line)
-        } catch {
-            // Yay! It failed again!
-        }
     }
 
     private func assertRoundTripJSON(
