+++ Tests/ExperimentalTableDrivenSwiftProtobufTests/Test_Map.swift
@@ -17,7 +17,7 @@ import Foundation
 import SwiftProtobuf
 import XCTest
 
-final class Test_Map: XCTestCase {
+final class Test_Map: XCTestCase, PBTestHelpers {
     typealias MessageTestType = SwiftProtoTesting_TestMap
 
     func testEmpty() {
@@ -1111,338 +1111,5 @@ final class Test_Map: XCTestCase {
         } catch let e {
             XCTFail("Serialization failed: \(e)", file: file, line: line)
         }
-    }
-
-    func assertDecodeSucceeds(
-        _ bytes: [UInt8],
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) -> Bool
-    ) {
-        do {
-            let decoded = try MessageTestType(serializedBytes: bytes)
-            XCTAssert(check(decoded), "Condition failed for decode", file: file, line: line)
-
-            do {
-                let encoded: [UInt8] = try decoded.serializedBytes()
-                do {
-                    let redecoded = try MessageTestType(serializedBytes: encoded)
-                    XCTAssert(check(redecoded), "Condition failed for redecoded", file: file, line: line)
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch let e {
-                    XCTFail("Failed to redecode: \(e)", file: file, line: line)
-                }
-            } catch let e {
-                XCTFail("Failed to encode: \(e)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Failed to decode: \(e)", file: file, line: line)
-        }
-
-        do {
-            // Make sure unknown fields are preserved by empty message decode/encode
-            let empty = try SwiftProtoTesting_TestEmptyMessage(serializedBytes: bytes)
-            do {
-                let newBytes: [UInt8] = try empty.serializedBytes()
-                XCTAssertEqual(
-                    bytes,
-                    newBytes,
-                    "Empty decode/recode did not match; \(bytes) != \(newBytes)",
-                    file: file,
-                    line: line
-                )
-            } catch let e {
-                XCTFail("Reserializing empty threw an error: \(e)", file: file, line: line)
-            }
-        } catch {
-            XCTFail("Empty decoding threw an error: \(error)", file: file, line: line)
-        }
-    }
-
-    func assertDecodeSucceeds(
-        inputBytes bytes: [UInt8],
-        recodedBytes: [UInt8],
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) -> Bool
-    ) {
-        do {
-            let decoded = try MessageTestType(serializedBytes: bytes)
-            XCTAssert(check(decoded), "Condition failed for \(decoded)", file: file, line: line)
-
-            do {
-                var options = BinaryEncodingOptions()
-                options.useDeterministicOrdering = true
-                let encoded: [UInt8] = try decoded.serializedBytes(options: options)
-                XCTAssertEqual(
-                    recodedBytes,
-                    encoded,
-                    "Didn't recode as expected: \(encoded) expected: \(recodedBytes)",
-                    file: file,
-                    line: line
-                )
-                do {
-                    let redecoded = try MessageTestType(serializedBytes: encoded)
-                    XCTAssert(check(redecoded), "Condition failed for redecoded", file: file, line: line)
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch let e {
-                    XCTFail("Failed to redecode: \(e)", file: file, line: line)
-                }
-            } catch let e {
-                XCTFail("Failed to encode: \(e)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Failed to decode: \(e)", file: file, line: line)
-        }
-    }
-
-    /// Verify the preferred encoding/decoding of a particular object.
-    /// This uses the provided block to initialize the object, then:
-    /// * Encodes the object and checks that the result is the expected result
-    /// * Decodes it again and verifies that the round-trip gives an equal object
-    func assertTextFormatEncode(
-        _ expected: String,
-        extensions: (any ExtensionMap)? = nil,
-        file: StaticString = #file,
-        line: UInt = #line,
-        configure: (inout MessageTestType) -> Void
-    ) {
-        let empty = MessageTestType()
-        var configured = empty
-        configure(&configured)
-        XCTAssert(configured != empty, "Object should not be equal to empty object", file: file, line: line)
-        let encoded = configured.textFormatString()
-
-        XCTAssertEqual(expected, encoded, "Did not encode correctly", file: file, line: line)
-        do {
-            let decoded = try MessageTestType(textFormatString: encoded, extensions: extensions)
-            XCTAssert(
-                decoded == configured,
-                "Encode/decode cycle should generate equal object: \(decoded) != \(configured)",
-                file: file,
-                line: line
-            )
-        } catch {
-            XCTFail(
-                "Encode/decode cycle should not throw error but got \(error) while decoding \(encoded)",
-                file: file,
-                line: line
-            )
-        }
-    }
-
-    func assertTextFormatDecodeSucceeds(
-        _ text: String,
-        options: TextFormatDecodingOptions = TextFormatDecodingOptions(),
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) throws -> Bool
-    ) {
-        do {
-            let decoded: MessageTestType = try MessageTestType(textFormatString: text, options: options)
-            do {
-                let r = try check(decoded)
-                XCTAssert(r, "Condition failed for \(decoded)", file: file, line: line)
-            } catch let e {
-                XCTFail("Object check failed: \(e)")
-            }
-            let encoded = decoded.textFormatString()
-            do {
-                let redecoded = try MessageTestType(textFormatString: encoded)
-                do {
-                    let r = try check(redecoded)
-                    XCTAssert(r, "Condition failed for redecoded \(redecoded)", file: file, line: line)
-                } catch let e {
-                    XCTFail("Object check failed for redecoded: \(e)\n   \(redecoded)")
-                }
-                XCTAssertEqual(decoded, redecoded, file: file, line: line)
-            } catch {
-                XCTFail("Swift should have recoded/redecoded without error: \(encoded)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Swift should have decoded without error but got \(e) decoding: \(text)", file: file, line: line)
-            return
-        }
-    }
-
-    func assertTextFormatDecodeFails(_ text: String, file: StaticString = #file, line: UInt = #line) {
-        do {
-            let _ = try MessageTestType(textFormatString: text)
-            XCTFail("Swift decode should have failed: \(text)", file: file, line: line)
-        } catch {
-            // Yay! It failed!
-        }
-    }
-
-    func assertJSONEncode(
-        _ expected: String,
-        extensions: any ExtensionMap = SimpleExtensionMap(),
-        encodingOptions: JSONEncodingOptions = .init(),
-        file: StaticString = #file,
-        line: UInt = #line,
-        configure: (inout MessageTestType) -> Void
-    ) {
-        let empty = MessageTestType()
-        var configured = empty
-        configure(&configured)
-        XCTAssert(configured != empty, "Object should not be equal to empty object", file: file, line: line)
-        do {
-            let encoded = try configured.jsonString(options: encodingOptions)
-            XCTAssert(
-                expected == encoded,
-                "Did not encode correctly: got \(encoded) but expected \(expected)",
-                file: file,
-                line: line
-            )
-            do {
-                let decoded = try MessageTestType(jsonString: encoded, extensions: extensions)
-                XCTAssert(
-                    decoded == configured,
-                    "Encode/decode cycle should generate equal object: \(decoded) != \(configured)",
-                    file: file,
-                    line: line
-                )
-            } catch {
-                XCTFail(
-                    "Encode/decode cycle should not throw error decoding: \(encoded), but it threw \(error)",
-                    file: file,
-                    line: line
-                )
-            }
-        } catch let e {
-            XCTFail("Failed to serialize JSON: \(e)\n    \(configured)", file: file, line: line)
-        }
-
-        do {
-            let encodedData: [UInt8] = try configured.jsonUTF8Bytes(options: encodingOptions)
-            let encodedOptString = String(bytes: encodedData, encoding: String.Encoding.utf8)
-            XCTAssertNotNil(encodedOptString)
-            let encodedString = encodedOptString!
-            XCTAssert(
-                expected == encodedString,
-                "Did not encode correctly: got \(encodedString)",
-                file: file,
-                line: line
-            )
-            do {
-                let decoded = try MessageTestType(jsonUTF8Bytes: encodedData, extensions: extensions)
-                XCTAssert(
-                    decoded == configured,
-                    "Encode/decode cycle should generate equal object: \(decoded) != \(configured)",
-                    file: file,
-                    line: line
-                )
-            } catch {
-                XCTFail(
-                    "Encode/decode cycle should not throw error decoding: \(encodedString), but it threw \(error)",
-                    file: file,
-                    line: line
-                )
-            }
-        } catch let e {
-            XCTFail("Failed to serialize JSON: \(e)\n    \(configured)", file: file, line: line)
-        }
-    }
-
-    func assertJSONDecodeSucceeds(
-        _ json: String,
-        options: JSONDecodingOptions = JSONDecodingOptions(),
-        extensions: any ExtensionMap = SimpleExtensionMap(),
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) -> Bool
-    ) {
-        do {
-            let decoded: MessageTestType = try MessageTestType(
-                jsonString: json,
-                extensions: extensions,
-                options: options
-            )
-            XCTAssert(check(decoded), "Condition failed for \(decoded)", file: file, line: line)
-
-            do {
-                let encoded = try decoded.jsonString()
-                do {
-                    let redecoded = try MessageTestType(jsonString: encoded, extensions: extensions, options: options)
-                    XCTAssert(
-                        check(redecoded),
-                        "Condition failed for redecoded \(redecoded) from \(encoded)",
-                        file: file,
-                        line: line
-                    )
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch {
-                    XCTFail("Swift should have recoded/redecoded without error: \(encoded)", file: file, line: line)
-                }
-            } catch let e {
-                XCTFail("Swift should have recoded without error but got \(e)\n    \(decoded)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Swift should have decoded without error but got \(e): \(json)", file: file, line: line)
-            return
-        }
-
-        do {
-            let jsonData = json.data(using: String.Encoding.utf8)!
-            let decoded: MessageTestType = try MessageTestType(
-                jsonUTF8Bytes: jsonData,
-                extensions: extensions,
-                options: options
-            )
-            XCTAssert(check(decoded), "Condition failed for \(decoded) from binary \(json)", file: file, line: line)
-
-            do {
-                let encoded: [UInt8] = try decoded.jsonUTF8Bytes()
-                let encodedString = String(decoding: encoded, as: UTF8.self)
-                do {
-                    let redecoded = try MessageTestType(
-                        jsonUTF8Bytes: encoded,
-                        extensions: extensions,
-                        options: options
-                    )
-                    XCTAssert(
-                        check(redecoded),
-                        "Condition failed for redecoded \(redecoded) from binary \(encodedString)",
-                        file: file,
-                        line: line
-                    )
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch {
-                    XCTFail(
-                        "Swift should have recoded/redecoded without error: \(encodedString)",
-                        file: file,
-                        line: line
-                    )
-                }
-            } catch let e {
-                XCTFail("Swift should have recoded without error but got \(e)\n    \(decoded)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Swift should have decoded without error but got \(e): \(json)", file: file, line: line)
-            return
-        }
-    }
-
-    func assertJSONDecodeFails(
-        _ json: String,
-        extensions: any ExtensionMap = SimpleExtensionMap(),
-        options: JSONDecodingOptions = JSONDecodingOptions(),
-        file: StaticString = #file,
-        line: UInt = #line
-    ) {
-        //        do {
-        //            let _ = try MessageTestType(jsonString: json, extensions: extensions, options: options)
-        //            XCTFail("Swift decode should have failed: \(json)", file: file, line: line)
-        //        } catch {
-        //            // Yay! It failed!
-        //        }
-        //
-        //        let jsonData = json.data(using: String.Encoding.utf8)!
-        //        do {
-        //            let _ = try MessageTestType(jsonUTF8Bytes: jsonData, extensions: extensions, options: options)
-        //            XCTFail("Swift decode should have failed for binary: \(json)", file: file, line: line)
-        //        } catch {
-        //            // Yay! It failed again!
-        //        }
     }
 }
