+++ Tests/ExperimentalTableDrivenSwiftProtobufTests/Test_Packed.swift
@@ -17,7 +17,7 @@ import Foundation
 import SwiftProtobuf
 import XCTest
 
-final class Test_Packed: XCTestCase {
+final class Test_Packed: XCTestCase, PBTestHelpers {
     typealias MessageTestType = SwiftProtoTesting_TestPackedTypes
 
     func testEncoding_packedInt32() {
@@ -536,186 +536,5 @@ final class Test_Packed: XCTestCase {
         assertTextFormatDecodeFails("packed_int32: [1,]\n")
         assertTextFormatDecodeFails("packed_int32: [1,2\n")
         assertTextFormatDecodeFails("packed_int32: [1,2,]\n")
-    }
-
-    func assertEncode(
-        _ expected: [UInt8],
-        file: StaticString = #file,
-        line: UInt = #line,
-        configure: (inout MessageTestType) -> Void
-    ) {
-        let empty = MessageTestType()
-        var configured = empty
-        configure(&configured)
-        XCTAssert(configured != empty, "Object should not be equal to empty object", file: file, line: line)
-        do {
-            let encoded: [UInt8] = try configured.serializedBytes()
-            XCTAssert(
-                expected == encoded,
-                "Did not encode correctly: got \(encoded)",
-                file: file,
-                line: line
-            )
-            do {
-                let decoded = try MessageTestType(serializedBytes: encoded)
-                XCTAssert(
-                    decoded == configured,
-                    "Encode/decode cycle should generate equal object",
-                    file: file,
-                    line: line
-                )
-            } catch let e {
-                XCTFail("Failed to decode protobuf: \(e)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Failed to encode: \(e)", file: file, line: line)
-        }
-    }
-
-    func assertDecodeSucceeds(
-        _ bytes: [UInt8],
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) -> Bool
-    ) {
-        do {
-            let decoded = try MessageTestType(serializedBytes: bytes)
-            XCTAssert(check(decoded), "Condition failed for decode", file: file, line: line)
-
-            do {
-                let encoded: [UInt8] = try decoded.serializedBytes()
-                do {
-                    let redecoded = try MessageTestType(serializedBytes: encoded)
-                    XCTAssert(check(redecoded), "Condition failed for redecoded", file: file, line: line)
-                    XCTAssertEqual(decoded, redecoded, file: file, line: line)
-                } catch let e {
-                    XCTFail("Failed to redecode: \(e)", file: file, line: line)
-                }
-            } catch let e {
-                XCTFail("Failed to encode: \(e)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Failed to decode: \(e)", file: file, line: line)
-        }
-
-        do {
-            // Make sure unknown fields are preserved by empty message decode/encode
-            let empty = try SwiftProtoTesting_TestEmptyMessage(serializedBytes: bytes)
-            do {
-                let newBytes: [UInt8] = try empty.serializedBytes()
-                XCTAssertEqual(bytes, newBytes, "Empty decode/recode did not match; \(bytes) != \(newBytes)", file: file, line: line)
-            } catch let e {
-                XCTFail("Reserializing empty threw an error: \(e)", file: file, line: line)
-            }
-        } catch {
-            XCTFail("Empty decoding threw an error: \(error)", file: file, line: line)
-        }
-    }
-
-    func assertDecodeFails(_ bytes: [UInt8], file: StaticString = #file, line: UInt = #line) {
-        do {
-            let _ = try MessageTestType(serializedBytes: bytes)
-            XCTFail("Swift decode should have failed: \(bytes)", file: file, line: line)
-        } catch {
-            // Yay!  It failed!
-        }
-
-    }
-
-    // Helper to check that decode succeeds by the data ended up in unknown fields.
-    // Supports an optional `check` to do additional validation.
-    func assertDecodesAsUnknownFields(
-        _ bytes: [UInt8],
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: ((MessageTestType) -> Bool)? = nil
-    ) {
-        assertDecodeSucceeds(bytes, file: file, line: line) {
-            if $0.unknownFields.data != Data(bytes) {
-                return false
-            }
-            if let check = check {
-                return check($0)
-            }
-            return true
-        }
-    }
-
-    /// Verify the preferred encoding/decoding of a particular object.
-    /// This uses the provided block to initialize the object, then:
-    /// * Encodes the object and checks that the result is the expected result
-    /// * Decodes it again and verifies that the round-trip gives an equal object
-    func assertTextFormatEncode(
-        _ expected: String,
-        extensions: (any ExtensionMap)? = nil,
-        file: StaticString = #file,
-        line: UInt = #line,
-        configure: (inout MessageTestType) -> Void
-    ) {
-        let empty = MessageTestType()
-        var configured = empty
-        configure(&configured)
-        XCTAssert(configured != empty, "Object should not be equal to empty object", file: file, line: line)
-        let encoded = configured.textFormatString()
-
-        XCTAssertEqual(expected, encoded, "Did not encode correctly", file: file, line: line)
-        do {
-            let decoded = try MessageTestType(textFormatString: encoded, extensions: extensions)
-            XCTAssert(
-                decoded == configured,
-                "Encode/decode cycle should generate equal object: \(decoded) != \(configured)",
-                file: file,
-                line: line
-            )
-        } catch {
-            XCTFail(
-                "Encode/decode cycle should not throw error but got \(error) while decoding \(encoded)",
-                file: file,
-                line: line
-            )
-        }
-    }
-
-    func assertTextFormatDecodeSucceeds(
-        _ text: String,
-        options: TextFormatDecodingOptions = TextFormatDecodingOptions(),
-        file: StaticString = #file,
-        line: UInt = #line,
-        check: (MessageTestType) throws -> Bool
-    ) {
-        do {
-            let decoded: MessageTestType = try MessageTestType(textFormatString: text, options: options)
-            do {
-                let r = try check(decoded)
-                XCTAssert(r, "Condition failed for \(decoded)", file: file, line: line)
-            } catch let e {
-                XCTFail("Object check failed: \(e)")
-            }
-            let encoded = decoded.textFormatString()
-            do {
-                let redecoded = try MessageTestType(textFormatString: encoded)
-                do {
-                    let r = try check(redecoded)
-                    XCTAssert(r, "Condition failed for redecoded \(redecoded)", file: file, line: line)
-                } catch let e {
-                    XCTFail("Object check failed for redecoded: \(e)\n   \(redecoded)")
-                }
-                XCTAssertEqual(decoded, redecoded, file: file, line: line)
-            } catch {
-                XCTFail("Swift should have recoded/redecoded without error: \(encoded)", file: file, line: line)
-            }
-        } catch let e {
-            XCTFail("Swift should have decoded without error but got \(e) decoding: \(text)", file: file, line: line)
-            return
-        }
-    }
-
-    func assertTextFormatDecodeFails(_ text: String, file: StaticString = #file, line: UInt = #line) {
-        do {
-            let _ = try MessageTestType(textFormatString: text)
-            XCTFail("Swift decode should have failed: \(text)", file: file, line: line)
-        } catch {
-            // Yay! It failed!
-        }
     }
 }
